Create ArrayDS1
In this code file you will learn about some Data structure of Array in
java. 
You are going to learn about 1-D  ,multi-dimensional array, sum of an Array, search in an array, search max element in an Array and also how to print array with the help of for loop and for-each loop.